backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.121.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
